refactor: move components to voting folder
